move add action form to another page
