+++ frontend/src/components/banner/Main.tsx
@@ -5,7 +5,7 @@ interface BannerProps {
 const Main = ({ heading }: BannerProps) => {
   return (
     <div className="banner">
-      <div className="container pt-32 pb-12 md:pt-36 md:pb-20">
+      <div className="container pt-40 md:pt-28 pb-12 md:pb-20">
         <Pill text={heading} />
       </div>
     </div>
