+++ frontend/src/styles/main.scss
@@ -100,6 +100,8 @@ select {
     top: 0;
     left: -25%;
     background-image: url(/images/arc-bg.png);
+    overflow: hidden;
+    position: relative;
     @screen md {
       top: 360px;
       left: 50%;
